% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ffpemodel}
\alias{ffpemodel}
\title{relationship between average tpm and ffpe artifacts based on a gam model}
\format{
a \code{\link[gamObject]{gamObject}} object
}
\usage{
ffpemodel
}
\description{
The model was learnt using seven paired FFPE-FFzn samples from BBD patients
}
\keyword{datasets}
