% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{gsea_rwplot}
\alias{gsea_rwplot}
\title{GSEA random-walk plot}
\usage{
gsea_rwplot(gseares_path, gsname, class_name, metric_range = NULL)
}
\arguments{
\item{gseares_path}{path to GSEA output.}

\item{gsname}{a character value indicating the gene-set name to generate random-walk plot.}

\item{class_name}{a character value indicating the class of the gene-set, e.g. "GO".}

\item{metric_range}{optional range of the ranking metric.}
}
\value{
a random-walk plot for selected gene-set
}
\description{
Generate high-quality GSEA random-walk figures.
}
\details{
this function does not support output from GSEA \href{https://github.com/GSEA-MSigDB/GSEA_R}{R implementation}. Scripts initially implemented by Thomas Kuilman.
}
