% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{gsea_sumplot}
\alias{gsea_sumplot}
\title{Summary plot of gene set enrichment analysis}
\usage{
gsea_sumplot(
  gseares_path,
  pos_sel,
  neg_sel,
  pvalflag = TRUE,
  interactive = FALSE
)
}
\arguments{
\item{gseares_path}{path to GSEA output.}

\item{pos_sel}{a character vector of upregulated gene-set names.}

\item{neg_sel}{a character vector of downregulated gene-set names.}

\item{pvalflag}{a logical value indicating whether to use adjusted p value in the plot. Default to TRUE.}

\item{interactive}{a logical value indicating whether to generate an interactive plot. Default to FALSE.}
}
\description{
Summary plot of selected up/down regulated gene-sets for gene set enrichment analysis.
}
\details{
this function does not support output from GSEA \href{https://github.com/GSEA-MSigDB/GSEA_R}{R implementation}
}
