% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{prepare_rnk}
\alias{prepare_rnk}
\title{Prepare .rnk file for GSEA preranked analysis}
\usage{
prepare_rnk(teststats, outputfile, replace = FALSE)
}
\arguments{
\item{teststats}{a data.frame containing differential expression results with five columns: "Gene name", "log2 fold change", "log2 average expression",
"p value", "adjusted p value". The second element of the output from function \code{\link{run_de}}.}

\item{outputfile}{full path including file name to .rnk file.}

\item{replace}{a logical value indicating whether to replace the output file if it already exists. Default to FALSE.}
}
\value{
No return value
}
\description{
Prepare .rnk file for GSEA preranked analysis
}
