% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{prop_barplot}
\alias{prop_barplot}
\title{Bar plot of cell type proportions across samples}
\usage{
prop_barplot(prop, sort = TRUE, interactive = FALSE)
}
\arguments{
\item{prop}{a matrix or data.frame of cell proportion values with rows representing cell types, columns representing samples.}

\item{sort}{a logical value indicating whether to sort the samples based on cell type with highest median cell proportion across samples. Default to TRUE.}

\item{interactive}{a logical value indicating whether to generate interactive plot. Default to FALSE.}
}
\description{
Bar plot of cell type proportions across samples
}
