% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refdata}
\alias{refdata}
\title{Integrated scRNA-seq reference for breast tissue (subset)}
\format{
`refdata`
a \code{\link[Seurat]{Seurat-class}} object with several metadata included:
\describe{
 \item{cellid}{cell barcodes}
 \item{celltypes}{annotated cell types}
 \item{subjectid}{subject id}
 \item{cohort}{data cohort, either "komentissuebank" or "gtex"}
}
}
\source{
<https://gtexportal.org/home/singleCellOverviewPage>
<A single-cell atlas of the healthy breast tissues reveals clinically relevant clusters of breast epithelial cells>
}
\usage{
refdata
}
\description{
Integrated scRNA-seq reference using random subset of two scRNA-seq datasets for breast tissue. One sample is from Komen tissue bank. 
The other sample is from Gtex snRNA-seq data.
}
\keyword{datasets}
