% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refdata_brain}
\alias{refdata_brain}
\title{Reference scRNA-seq data for mouse cortical cell taxonomy from the Allen Institute (subset)}
\format{
a \code{\link[Seurat]{Seurat-class}} object with several metadata included:
\describe{
 \item{cellid}{cell barcodes}
 \item{celltypes}{annotated cell types}
 \item{subjectid}{subject id}
}
}
\source{
\url{https://www.nature.com/articles/nn.4216}
}
\usage{
refdata_brain
}
\description{
The data was generated using SMART-Seq2 protocol. 20 cells were randomly selected for each cell type. 15,000 highly variable
genes were selected.
}
\keyword{datasets}
