% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCdeconR-package.R
\docType{package}
\name{SCdeconR-package}
\alias{SCdeconR}
\alias{SCdeconR-package}
\title{SCdeconR: Deconvolution of Bulk RNA-Seq Data using Single-Cell RNA-Seq Data as Reference}
\description{
Streamlined workflow from deconvolution of bulk RNA-seq data to downstream differential expression and gene-set enrichment analysis. Provide various visualization functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Liuy12/SCdeconR/}
  \item \url{https://liuy12.github.io/SCdeconR/}
}

}
\author{
\strong{Maintainer}: Yuanhang Liu \email{liu.yuanhang@mayo.edu}

}
\keyword{internal}
