% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transformation}
\alias{transformation}
\title{Transformation of gene expression data}
\usage{
transformation(matrix, option)
}
\arguments{
\item{matrix}{a matrix-like objector of gene expression values with rows representing genes, columns representing samples or cells}

\item{option}{character value specifying the transformation method to use. Has to be one of "none", "log", "sqrt", "vst".}
}
\value{
a matrix-like object with the same dimension of input object after data transformation.
}
\description{
Methods to use for data transformation.
}
\details{
refer to \code{\link{scdecon}} for more details.
}
